requirements.txt
appdirs==1.4.3
Django==1.11
django-crispy-forms==1.6.1
django-registration-redux==1.5
gunicorn==19.7.1
olefile==0.44
packaging==16.8
Pillow==4.1.0
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
